refactor: Remove "components" directory
