+++ src/editor/index.ts
@@ -1,2 +1,3 @@
 export { EditorProps } from './editor'
 export { createPlugins } from './plugins'
+export { Layout } from './layout'
