+++ src/pages/index.tsx
@@ -2,9 +2,8 @@ import { GetServerSideProps } from 'next'
 import dynamic from 'next/dynamic'
 
 import { kitchenSink } from '../kitchen-sink'
-import { Layout } from '../components'
 import { migrate, emptyDocument } from '../storage-format'
-import { createPlugins } from '../editor'
+import { createPlugins, Layout } from '../editor'
 import { getJsonBody } from '../utils/get-json-body'
 import { Renderer } from '@edtr-io/renderer'
 import type { EditorProps } from '../editor'
